feat: endpoint POST /api/depositos/nuevo + GET de depósitos
